task3_AI_RESUME_SCREENER/requirements.txt
langchain>=0.3.0
langchain-core>=0.3.0
langchain-groq>=1.0.0
python-dotenv>=1.0.1
langsmith>=0.3.0
